FIXME: arm64: dts: qcom: msm8916-samsung-fortuna3g: Enable I2C6
fortuna3g uses rt5033 charger. Proper drivers are yet to be
made for rt5033. It is connected to I2C6 bus on
gpio 22 and 23.

Enabling I2C6 allows disabling and enabling charger from
userspace. Disabling charger helps to increase battery
life in cases when device needs to be always on
and has to stay plugged in. Charger can be enabled when
battery drops below specific level.

This commit is to be replaced with bindings for rt5033
when proper driver becomes available.

Signed-off-by: Siddharth Manthan <siddharth.manthan@gmail.com>